Creating the Qrcode
New file and class named Pix, where it has a method called create_payment. It generates a simulated bank code. We create a copy_paste code. Based on this code we generate a qrcodem and we salve it in the folder static/img. After that it returns the generated id from the bank, and the path to the equivalent qrcode img