pix confirmation
Implementing pix_confirmation function, checking if the user sended both value and ban_paiment_id. Also checking if the payment (through bank_payment_id) exists and if the payment was not done yet. And also checking if the value gives matches the value in the db for that specific payment. If everything goes smooth, it turns payment.paid into True and it commits in the db and returns a message of confirmation to the user